requirements.txt
django==4.2.6
djangorestframework==3.14.0
phonenumbers==8.13.23
channels==4.0.0
daphne==4.0.0
pyotp==2.9.0
Pillow==10.1.0
django_otp==1.2.4
django-cors-headers==4.3.0